cleaning up some admin stuff

+++ Portfolio/Pages/Admin/ProjectLair.cshtml.cs
@@ -100,7 +100,8 @@ namespace Portfolio.Pages.Admin
         /// <returns> redirect in place </returns>
         public async Task<IActionResult> OnPostAddImage(IFormFile file)
         {
-            await _uploadService.AddProjectImage(file);
+            if (file != null)
+                await _uploadService.AddProjectImage(file);
             return Redirect("/Admin/ProjectLair");
         }
 
@@ -112,8 +113,11 @@ namespace Portfolio.Pages.Admin
         /// <returns> redirects in place </returns>
         public async Task<IActionResult> OnPostUpdateImage(IFormFile file)
         {
-            await _adminContext.DeleteBlobImage(Project.FileName);
-            await _uploadService.UpdateImage(file, Project.Id);
+            if (file != null)
+            {
+                await _adminContext.DeleteBlobImage(Project.FileName);
+                await _uploadService.UpdateImage(file, Project.Id);
+            }
 
             return Redirect("/Admin/ProjectLair");
         }
